web/pyproject.toml
[project]
name = "htmx-django"
version = "0.1.0"
description = "A Django project using HTMX"
readme = "README.md"

requires-python = ">=3.13,<3.14"
dependencies = [
    "Django==6.0",
    "django-htmx==1.27.0",
    "psycopg2==2.9.11",
    "pydantic==2.12.5",
    "django-allauth==65.13.1",
    "gunicorn==23.0.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.9",
    "pytest==9.0.2",
    "django-debug-toolbar==6.1.0",
    "pytest-django==4.11.1",
    "factory-boy==3.3.3",
    "djlint==1.36.4",
    "django-stubs==5.2.8",
    "pyrefly==0.45.2",
    "taskipy==1.14.1",
    "pytest-randomly==4.0.1",
]

[tool.uv]
link-mode = "copy"

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pyrefly]
python-version = "3.13.0"
ignore-missing-imports = ["debug_toolbar.*"]
search_path = [
    "."
]


[tool.ruff.lint]
ignore = [
]

select = [
   "I",
]

[tool.taskipy.tasks]
check = """
uv run ruff check . \
    && uv run pyrefly check --summarize-errors \
    && uv run pyrefly check --remove-unused-ignores \
    && uv run djlint templates --extension html
"""

fix = """
uv run ruff check . --fix \
    && uv run ruff format . \
    && uv run djlint templates --extension html --reformat
"""

test = """
uv run pytest
"""

migration = """
uv run python manage.py makemigrations \
    && uv run python manage.py migrate
"""

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
